index-building/requirements.txt
google-genai
google-cloud-aiplatform
python-dotenv
pydantic
pytest